Using nighttime lights for economic inference needs cleaning of the data.

NighttimeLights.linear_interpolation — Method

Uses linear interpolation to fill for missing values. Missing and NaN are used interchangeably.

Example:

julia> x = rand(1:10.0, 10)
julia> x[5] = NaN
julia> linear_interpolation(x)

NighttimeLights.outlier_mask — Method

There are extremely high values in the data due to fires, gas flare etc. You may find some values even greater than the aggregate radiance of large cities. Such pixels also have high standard deviation. These pixels may not be of importantance from the point of view of measureming prosperity. The outliermask function generates a mask of pixels with standard deviation less that the 99.9th percentile. Essentially, this function can be used to removed top 1 percent of pixels by standard deviation. A mask can be provided to the function, so that it calculates the percentile based on the lit pixel of the mask. For example, if the datacube is a box around India and the mask is the polygon mask of India, the outliermask function will calculate the 99th percentile of the standard deviation of the pixels inside India's boundary.

outlier_mask(datacube, mask)

NighttimeLights.outlier_ts — Method

The time series of a pixel may show a few outliers, but as a whole the pixel may be of importantance in measuring economic activity. The outlier_ts function uses replaces the outlier observations with interpolated values. This is done using the tsclean function the forecast package of R.

sample_timeseries = datacube[1, 2, :] # The time series of pixel [1,2]
outlier_ts(sample_timeseries)

There are negative observations in the data as a consequence of NOAA's cleaning procedures. Many researchers replace them with 0. It can be done using the following:

datacube = replace!(x -> x < 0 ? 0 : x, datacube) # replace all values below 0 with 0

These values can be replaced with NaN and interpolated.

datacube = replace!(x -> x < 0 ? NaN : x, datacube) # replace all values below 0 with NaN

NighttimeLights.background_noise_mask — Function

Pixels with no economic activity may show some light due to background noise. These pixels could be in forests, oceans, deserts etc. The backgroundnoisemask function generates a background moise mask such that those pixels which are considered dark are marked as 0 and those considered lit are marked as 1. The function uses the datacubes of radiance and clouds to generate annual image of the last year the data. The function considers all the pixels below a provided threshold as dark and remaining to be lit.