Using nighttime lights for economic inference needs cleaning of the data.

There are negative observations in the data as a consequence of NOAA's cleaning procedures. Many researchers replace them with 0. It can be done using the following:

datacube = replace!(x -> x < 0 ? 0 : x, datacube) # replace all values below 0 with 0

These values can be replaced with missing and interpolated.

datacube = replace!(x -> x < 0 ? missing : x, datacube) # replace all values below 0 with missing

Missing docstring.

Missing docstring for mark_missing(radiance::Array{T, 2}, clouds) where T <: Any. Check Documenter's build log for details.

Missing docstring.

Missing docstring for mark_missing(radiance_datacube::Array{T, 3}, clouds_datacube) where T <: Any. Check Documenter's build log for details.

NighttimeLights.linear_interpolation — Method

Uses linear interpolation to fill for missing values.

Example:

x = rand(1:10.0, 10)
x[5] = missing
linear_interpolation(x)

NighttimeLights.outlier_mask — Method

There are extremely high values in the data due to fires, gas flare etc. You may find some values even greater than the aggregate radiance of large cities. Such pixels also have high standard deviation. These pixels may not be of importantance from the point of view of measureming prosperity. The outlier_mask function generates a mask of pixels with standard deviation less that the 99.9th percentile. Essentially, this function can be used to removed top 1 percent of pixels by standard deviation. A mask can be provided to the function, so that it calculates the percentile based on the lit pixel of the mask. For example, if the datacube is a box around India and the mask is the polygon mask of India, the outlier_mask function will calculate the 99th percentile of the standard deviation of the pixels inside India's boundary.

outlier_mask(datacube, mask)

Missing docstring.

Missing docstring for outlier_ts(timeseries, window_size = 5, n_sigmas = 3)background_noise_mask(datacube=radiance_datacube, clouds=clouds_datacube, th=0.4). Check Documenter's build log for details.

Missing docstring.

Missing docstring for bias_correction(radiance::Array{T, 1}, clouds) where T <:Real. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bias_correction(radiance_datacube::Array{T, 3}, clouds_datacube, mask=ones(Int8, (size(radiance_datacube)[1],size(radiance_datacube)[2]))) where T <:Real. Check Documenter's build log for details.

Missing docstring.

Missing docstring for conventonal_cleaning(radiance_datacube, clouds_datacube). Check Documenter's build log for details.

NighttimeLights.PatnaikSTT2021 — Method

The PatnaikSTT2021 function performs all the steps of the new cleaning procedure described in But clouds got in my way: Bias and bias correction of VIIRS nighttime lights data in the presence of clouds, Ayush Patnaik, Ajay Shah, Anshul Tayal, Susan Thomas as conventional cleaning.

Example

radiance_datacube = rand(1:1000, 10, 10, 10)
clouds_datacube = rand(1:1000, 10, 10, 10)
PatnaikSTT2021(radiance_datacube, clouds_datacube)